GenDL_Folk-main/requirements_ubuntu.txt
absl-py==2.1.0
amdsmi @ file:///opt/rocm-6.0.0/share/amd_smi
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
clang==17.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.13.3
flatbuffers==24.3.25
future==0.18.2
gast==0.4.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
install==1.3.5
jax==0.4.26
jeepney==0.7.1
keras==2.12.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-rocm==2.12.1.600
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.5.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
